Module 48 Project/Ner project/customized-chatbot-proj/requirements.txt
groq 
langchain_community 
qdrant_client 
langchain_groq 
langchain_qdrant 
pypdfium2
python-dotenv
flask
fastembed
sentence-transformers